backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.10.0
easyocr==1.7.1
PyPDF2==3.0.1
spacy==3.7.2
nltk==3.8.1
transformers==4.37.2
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
httpx==0.26.0
pydantic[email]==2.5.3
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
cloudinary==1.38.0
fastapi-mail==1.4.1
fastapi-sso==0.7.0
